Refactor power screen into view, viewmodel, rows, format modules
